requirements.txt
Django==3.0.2
django-environ==0.4.5
psycopg2==2.8.4
strawberry-graphql==0.20.2
requests==2.22.0
